requirements.txt
jupyter
ipywidgets
boto3
strands-agents
strands-agents-tools
python-dotenv
ddgs
mcp
uv